pyproject.toml
[project]
name = "lims"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=23.1.0",
]
